apparently float32 was too low precision, caused some problems

+++ bostonBusMap/src/boston/Bus/Map/data/BusLocation.java
@@ -110,8 +110,8 @@ public class BusLocation implements Location {
 			String routeName, Directions directions, String routeTitle,
 			boolean disappearAfterRefresh, boolean showBusNumber,
 			int arrowTopDiff) {
-		this.latitude = latitude * Geometry.degreesToRadians;
-		this.longitude = longitude * Geometry.degreesToRadians;
+		this.latitude = (float) (latitude * Geometry.degreesToRadians);
+		this.longitude = (float) (longitude * Geometry.degreesToRadians);
 		this.latitudeAsDegrees = latitude;
 		this.longitudeAsDegrees = longitude;
 		this.id = id;
@@ -371,8 +371,8 @@ public class BusLocation implements Location {
 	}
 
 	public void movedTo(float latitudeAsDegrees, float longitudeAsDegrees) {
-		movedFrom(latitudeAsDegrees * Geometry.degreesToRadians,
-				longitudeAsDegrees * Geometry.degreesToRadians);
+		movedFrom(((float)(latitudeAsDegrees * Geometry.degreesToRadians)),
+				((float)(longitudeAsDegrees * Geometry.degreesToRadians)));
 
 		distanceFromLastX *= -1;
 		distanceFromLastY *= -1;
